Rapid-Rescue-main/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.0.1
PyJWT>=1.7.1,<3.0.0
python-dotenv==1.0.1
drf-yasg==1.21.7
requests==2.31.0
django-filter==23.5
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-rest-swagger==2.2.0
pydantic==2.6.1
djangorestframework-api-key==2.3.0
django-import-export==3.3.5 